Refactor: 메인 화면 통합, props 전달
